Move document.getElementsByTagName into setTimeout
This gives iframes the 500ms time to init. Until now I had some frames initially with about:blank as src and when setTimeout callback fired, it sent about:blanks to the Inspector although by that time src was already changed to actual URL.

This is not a "prefect fix" but this reduces cases where Ember Apps in iframes are not detected.